Be explicit that Fos32 is an unsigned value.

+++ src/main/java/com/sibilantsolutions/grison/driver/foscam/dto/AudioDataTextDto.java
@@ -55,7 +55,7 @@ public abstract class AudioDataTextDto implements FoscamTextDto {
         public AudioDataTextDto build() {
             final AudioDataTextDto dto = autoBuild();
 
-            checkArgument(dto.dataLength().value() == dto.data().length, "data len expected=%s actual=%s", dto.dataLength(), dto.data().length);
+            checkArgument(dto.dataLength().value().longValue() == dto.data().length, "data len expected=%s actual=%s", dto.dataLength(), dto.data().length);
 
             return dto;
         }
